+++ src/example/entity/CreateEntityExample.vue
@@ -85,8 +85,7 @@
       const mapperOptions = {
         booleanLabels: {
           trueLabel: this.$t('form_boolean_true'), // $t is set via @molgenis/molgenis-i18n-js plugin
-          falseLabel: this.$t('form_boolean_false'),
-          nillLabel: this.$t('form_boolean_missing')
+          falseLabel: this.$t('form_boolean_false')
         },
         mapperMode: 'CREATE'
       }