+++ src/flow.types.js
@@ -43,10 +43,8 @@ export type MapperOptions = {
   mapperMode?: MapperMode,
   booleanLabels?: {
     trueLabel: string,
-    falseLabel: string,
-    nillLabel: string
+    falseLabel: string
   },
-  showNillableBooleanOption?: boolean,
   showNonVisibleAttributes?: boolean
 }
 
@@ -57,10 +55,8 @@ export type MapperSettings = {
   mapperMode: MapperMode,
   booleanLabels: {
     trueLabel: string,
-    falseLabel: string,
-    nillLabel: string
+    falseLabel: string
   },
-  showNillableBooleanOption: boolean,
   showNonVisibleAttributes: boolean
 }